+++ kernel/uid/uid_test.go
@@ -75,260 +75,6 @@ func (r *testErrorReader) Read(p []byte) (n int, err error) {
 	return 0, r.err
 }
 
-func TestIDValidator(t *testing.T) {
-	type condition struct {
-		typ kernel.EncodingType
-	}
-
-	type action struct {
-		shouldNil bool
-		valid     []string
-		invalid   []string
-	}
-
-	tb := testutil.NewTableBuilder[*condition, *action]()
-	tb.Name(t.Name())
-	cndValidType := tb.Condition("valid type", "input a valid encode type")
-	cndUnknown := tb.Condition("unknown type", "input a invalid encode type")
-	actCheckType := tb.Action("check type", "check the returned encode type")
-	table := tb.Build()
-
-	must := func(v []byte, err error) []byte {
-		if err != nil {
-			panic(err)
-		}
-		return v
-	}
-
-	gen := testutil.NewCase[*condition, *action]
-	testCases := []*testutil.Case[*condition, *action]{
-		gen(
-			"Base16",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base16,
-			},
-			&action{
-				valid: []string{
-					hex.EncodeToString(must(uid.NewID())),
-					hex.EncodeToString(must(uid.NewHostedID())),
-					"1234567890abcdefABCDEF0123456789abcdefABCDEF0123456789abcdef", // 60 chars.
-				},
-				invalid: []string{
-					"12345678901234567890123456789012345678901234567890123456789",   // 59 chars.
-					"1234567890123456789012345678901234567890123456789012345678901", // 61 chars.
-					"12345678901234567890123456789012345678901234567890123456789$",  // 60 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base32",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base32,
-			},
-			&action{
-				valid: []string{
-					base32.StdEncoding.EncodeToString(must(uid.NewID())),
-					base32.StdEncoding.EncodeToString(must(uid.NewHostedID())),
-					"234567ABCDEFGHIJKLMNOPQRSTUVWXYZ234567ABCDEFGHIJ", // 48 chars.
-				},
-				invalid: []string{
-					"234567ABCDEFGHIJKLMNOPQRSTUVWXYZ234567ABCDEFGHI",   // 47 chars.
-					"234567ABCDEFGHIJKLMNOPQRSTUVWXYZ234567ABCDEFGHIJK", // 49 chars.
-					"234567ABCDEFGHIJKLMNOPQRSTUVWXYZ234567ABCDEFGHI$",  // 48 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base32Hex",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base32Hex,
-			},
-			&action{
-				valid: []string{
-					base32.HexEncoding.EncodeToString(must(uid.NewID())),
-					base32.HexEncoding.EncodeToString(must(uid.NewHostedID())),
-					"1234567890ABCDEFGHIJKLMNOPQRSTUV1234567890ABCDEF", // 48 chars.
-				},
-				invalid: []string{
-					"1234567890ABCDEFGHIJKLMNOPQRSTUV1234567890ABCDE",   // 47 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUV1234567890ABCDEFG", // 49 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUV1234567890ABCDE$",  // 48 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base32Escaped",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base32Escaped,
-			},
-			&action{
-				valid: []string{
-					"1234567890BCDFGHJKLMNPQRSTUVWXYZ1234567890BCDFGH", // 48 chars.
-				},
-				invalid: []string{
-					"1234567890BCDFGHJKLMNPQRSTUVWXYZ1234567890BCDFG",   // 47 chars.
-					"1234567890BCDFGHJKLMNPQRSTUVWXYZ1234567890BCDFGHJ", // 49 chars.
-					"1234567890BCDFGHJKLMNPQRSTUVWXYZ1234567890BCDFG$",  // 48 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base32HexEscaped",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base32HexEscaped,
-			},
-			&action{
-				valid: []string{
-					"1234567890BCDFGHJKLMNPQRSTUV1234567890BCDFGHJKLM", // 48 chars.
-				},
-				invalid: []string{
-					"1234567890BCDFGHJKLMNPQRSTUV1234567890BCDFGHJKL",   // 47 chars.
-					"1234567890BCDFGHJKLMNPQRSTUV1234567890BCDFGHJKLMN", // 49 chars.
-					"1234567890BCDFGHJKLMNPQRSTUV1234567890BCDFGHJKL$",  // 48 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base64",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base64,
-			},
-			&action{
-				valid: []string{
-					base64.StdEncoding.EncodeToString(must(uid.NewID())),
-					base64.StdEncoding.EncodeToString(must(uid.NewHostedID())),
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/00", // 40 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/00", // 40 chars.
-				},
-				invalid: []string{
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/000", // 39 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/000", // 39 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/000", // 41 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/000", // 41 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/0$",  // 40 chars with invalid char.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/0$",  // 40 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base64Raw",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base64Raw,
-			},
-			&action{
-				valid: []string{
-					base64.RawStdEncoding.EncodeToString(must(uid.NewID())),
-					base64.RawStdEncoding.EncodeToString(must(uid.NewHostedID())),
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/00", // 40 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/00", // 40 chars.
-				},
-				invalid: []string{
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/000", // 39 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/000", // 39 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/000", // 41 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/000", // 41 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ+/0$",  // 40 chars with invalid char.
-					"1234567890abcdefghijklmnopqrstuvwxyz+/0$",  // 40 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base64URL",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base64URL,
-			},
-			&action{
-				valid: []string{
-					base64.URLEncoding.EncodeToString(must(uid.NewID())),
-					base64.URLEncoding.EncodeToString(must(uid.NewHostedID())),
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_00", // 40 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_00", // 40 chars.
-				},
-				invalid: []string{
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_000", // 39 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_000", // 39 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_000", // 41 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_000", // 41 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_0$",  // 40 chars with invalid char.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_0$",  // 40 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Base64RawURL",
-			[]string{cndValidType},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType_Base64RawURL,
-			},
-			&action{
-				valid: []string{
-					base64.RawURLEncoding.EncodeToString(must(uid.NewID())),
-					base64.RawURLEncoding.EncodeToString(must(uid.NewHostedID())),
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_00", // 40 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_00", // 40 chars.
-				},
-				invalid: []string{
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_000", // 39 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_000", // 39 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_000", // 41 chars.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_000", // 41 chars.
-					"1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ-_0$",  // 40 chars with invalid char.
-					"1234567890abcdefghijklmnopqrstuvwxyz-_0$",  // 40 chars with invalid char.
-				},
-			},
-		),
-		gen(
-			"Unknown",
-			[]string{cndUnknown},
-			[]string{actCheckType},
-			&condition{
-				typ: kernel.EncodingType(999),
-			},
-			&action{
-				shouldNil: true,
-			},
-		),
-	}
-
-	testutil.Register(table, testCases...)
-
-	for _, tt := range table.Entries() {
-		tt := tt
-		t.Run(tt.Name(), func(t *testing.T) {
-			reg := uid.Validator(tt.C().typ)
-			if tt.A().shouldNil {
-				testutil.Diff(t, (*regexp.Regexp)(nil), reg)
-				return
-			}
-			for _, v := range tt.A().valid {
-				t.Log("expect true", v)
-				testutil.Diff(t, true, reg.MatchString(v))
-			}
-			for _, v := range tt.A().invalid {
-				t.Log("expect false", v)
-				testutil.Diff(t, false, reg.MatchString(v))
-			}
-		})
-	}
-}
-
 func TestNewID(t *testing.T) {
 	type condition struct {
 		encodeFunc     func([]byte) string
@@ -463,7 +209,7 @@ func TestNewID(t *testing.T) {
 			}
 
 			eid := string(tt.C().encodeFunc(id))
-			pattern := uid.Validator(tt.C().typ)
+			pattern := Validator(tt.C().typ)
 			testutil.Diff(t, true, pattern.MatchString(eid))
 		})
 	}
@@ -603,8 +349,43 @@ func TestNewHostedID(t *testing.T) {
 			}
 
 			eid := string(tt.C().encodeFunc(id))
-			pattern := uid.Validator(tt.C().typ)
+			pattern := Validator(tt.C().typ)
 			testutil.Diff(t, true, pattern.MatchString(eid))
 		})
+	}
+}
+
+// Validator returns a regular expression for IDs.
+// Returned expression can be used for validating IDs
+// created by NewID() and NewHostedID().
+// nil will be returned if an unknown EncodeType was given.
+// Returned regular expressions are as follow.
+//   - Base16           : ^[0-9a-fA-F]{60}$
+//   - Base32           : ^[2-7A-Z]{48}$
+//   - Base32Hex        : ^[0-9A-V]{48}$
+//   - Base32Escaped    : ^[0-9B-DF-HJ-NP-TV-Z]{48}$
+//   - Base32HexEscaped : ^[0-9B-DF-HJ-NP-TV-Z]{48}$
+//   - Base64           : ^[0-9a-zA-Z+/]{40}$
+//   - Base64Raw        : ^[0-9a-zA-Z+/]{40}$
+//   - Base64URL        : ^[0-9a-zA-Z-_]{40}$
+//   - Base64RawURL     : ^[0-9a-zA-Z-_]{40}$
+func Validator(t kernel.EncodingType) *regexp.Regexp {
+	switch t {
+	case kernel.EncodingType_Base16:
+		return regexp.MustCompile(`^[0-9a-fA-F]{60}$`)
+	case kernel.EncodingType_Base32:
+		return regexp.MustCompile(`^[2-7A-Z]{48}$`)
+	case kernel.EncodingType_Base32Escaped:
+		return regexp.MustCompile(`^[0-9B-DF-HJ-NP-TV-Z]{48}$`)
+	case kernel.EncodingType_Base32Hex:
+		return regexp.MustCompile(`^[0-9A-V]{48}$`)
+	case kernel.EncodingType_Base32HexEscaped:
+		return regexp.MustCompile(`^[0-9B-DF-HJ-NP-TV-Z]{48}$`)
+	case kernel.EncodingType_Base64, kernel.EncodingType_Base64Raw:
+		return regexp.MustCompile(`^[0-9a-zA-Z+/]{40}$`)
+	case kernel.EncodingType_Base64URL, kernel.EncodingType_Base64RawURL:
+		return regexp.MustCompile(`^[0-9a-zA-Z-_]{40}$`)
+	default:
+		return nil
 	}
 }
